+++ WordPressKit/ReaderSiteServiceRemote.m
@@ -1,5 +1,4 @@
 #import "ReaderSiteServiceRemote.h"
-#import "RemoteReaderSite.h"
 #import "WPKit-Swift.h"
 @import NSObject_SafeExpectations;
 @import WordPressShared;